Fix RISC V specific makefiles not being run
Fix an erroneous ifeq statement at the top of the RISC V specific
makefiles which would stop them from being run when the TARGET was set
to anything but riscv32_mcu. Since the only way to get the makefile to
run was to set the target to mcu_riscv, the RISC V specific makefiles
were not getting run.