Add code tweaks for CV32E40P compatibility
Add some tweaks to the code guarded by preprocessor statements which
tweak some parts of the kernal and hello_world example to give
compatibility with the CV32E40P.

This was done to counter some known bugs in the RISC V GCC which use the
wrong version of the standard library.

They also provide print output to the hello_world program as the
DebugLog function used to provide output has not been implemented yet.