Add support for running conv unit tests on Spike
Add a conditional check for "spike" in TAGS. If it is found, all the
linking options enabling support for the Verilator model of the cv32e40p
are removed.
Spike requires programs to be compiled as if they were running utop the
Linux kernel. Removing these linking options causes GCC to default to
compilation for Linux Support.

Remove all unit test sources save for the conv_test source. This change
is temporary and should be reverted when development is complete.